converts data from client to json and save to database
Sometimes the server will fail to receive post request from client, currently I'm not sure why this is the case. When running on cmd servermain.py can also have a huge delay when trying to force stop with ctrl+c, not sure if this is how http.server module works or something is wrong with my implementation.